[CLEANUP] Move the shared configuration access to a trait
This reduces code duplication.
